refactor: split OnnxToHip.cpp monolith into per-operator files
Split the 2458-line OnnxToHip.cpp into 14 per-operator conversion files
and a shared OnnxToHipUtils.h header. This mirrors the HipToLLVM split
and eliminates the second single-file bottleneck in the operator
pipeline.

Each ONNX-to-HIP rewrite pattern is now in its own translation unit,
with a populate*ConversionPatterns() function called by convertComputeOps.
The shared header provides inline helpers (createEmptyTensor, getContextArg,
sanitizeForMlirIdentifier). Pass-internal code (ExternalizationState,
constant handling, metadata generation) stays in OnnxToHip.cpp.

No behavioral changes - pure mechanical code movement.
