+++ lib/Conversion/OnnxToHip/GatherConversion.cpp
@@ -1,0 +1,94 @@
+/*
+ * Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+#include "OnnxToHipUtils.h"
+
+namespace mlir {
+namespace hip {
+namespace {
+
+//===----------------------------------------------------------------------===//
+// ONNX Gather -> HIP Gather
+//===----------------------------------------------------------------------===//
+
+struct GatherToHip : public mlir::RewritePattern {
+  GatherToHip(mlir::MLIRContext *ctx)
+      : RewritePattern("onnx.Gather", /*benefit=*/1, ctx) {}
+
+  mlir::LogicalResult
+  matchAndRewrite(mlir::Operation *op,
+                  mlir::PatternRewriter &rewriter) const override {
+    auto ctxOrFailure = getContextArg(op, rewriter);
+    if (mlir::failed(ctxOrFailure))
+      return rewriter.notifyMatchFailure(op, "missing context argument");
+    mlir::Value context = *ctxOrFailure;
+
+    mlir::Location loc = op->getLoc();
+    mlir::Value data = op->getOperand(0);
+    mlir::Value indices = op->getOperand(1);
+
+    // Get axis attribute from ONNX Gather operation
+    int64_t axis = op->getAttrOfType<mlir::IntegerAttr>("axis").getSInt();
+    auto axisAttr = rewriter.getI64IntegerAttr(axis);
+
+    // Get result type
+    auto resultType =
+        mlir::cast<mlir::RankedTensorType>(op->getResult(0).getType());
+    auto dataType = mlir::cast<mlir::RankedTensorType>(data.getType());
+    auto indicesType = mlir::cast<mlir::RankedTensorType>(indices.getType());
+
+    // Normalize negative axis for dimension calculations only
+    int64_t normalizedAxis = axis < 0 ? axis + dataType.getRank() : axis;
+
+    // Create output tensor with dynamic shape support
+    // Output shape: [data[0:axis], indices.shape, data[axis+1:]]
+    llvm::SmallVector<mlir::Value> dynSizes;
+    int64_t outDimIdx = 0;
+
+    // Copy dimensions before axis from data
+    for (auto i : llvm::seq<int64_t>(0, normalizedAxis)) {
+      if (outDimIdx < resultType.getRank() &&
+          resultType.isDynamicDim(outDimIdx))
+        dynSizes.push_back(mlir::tensor::DimOp::create(rewriter, loc, data, i));
+      outDimIdx++;
+    }
+    // Copy all dimensions from indices
+    for (auto i : llvm::seq<int64_t>(0, indicesType.getRank())) {
+      if (outDimIdx < resultType.getRank() &&
+          resultType.isDynamicDim(outDimIdx))
+        dynSizes.push_back(
+            mlir::tensor::DimOp::create(rewriter, loc, indices, i));
+      outDimIdx++;
+    }
+    // Copy dimensions after axis from data
+    for (auto i : llvm::seq<int64_t>(normalizedAxis + 1, dataType.getRank())) {
+      if (outDimIdx < resultType.getRank() &&
+          resultType.isDynamicDim(outDimIdx))
+        dynSizes.push_back(mlir::tensor::DimOp::create(rewriter, loc, data, i));
+      outDimIdx++;
+    }
+
+    mlir::Value init =
+        mlir::tensor::EmptyOp::create(rewriter, loc, resultType.getShape(),
+                                      resultType.getElementType(), dynSizes);
+
+    // Create hip.gather operation
+    auto gatherOp = mlir::hip::GatherOp::create(
+        rewriter, loc, resultType, context, data, indices, init, axisAttr);
+
+    rewriter.replaceOp(op, gatherOp->getResult(0));
+    return mlir::success();
+  }
+};
+
+} // namespace
+
+void mlir::hip::populateGatherConversionPatterns(RewritePatternSet &patterns,
+                                                 MLIRContext *ctx) {
+  patterns.add<GatherToHip>(ctx);
+}
+
+} // namespace hip
+} // namespace mlir
