+++ lib/Conversion/OnnxToHip/OnnxToHipUtils.h
@@ -1,0 +1,146 @@
+/*
+ * Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
+ * Licensed under the MIT License.
+ */
+//===- OnnxToHipUtils.h - Shared helpers for ONNX-to-HIP patterns --------===//
+//
+// Shared utility functions and forward declarations used by per-operator
+// conversion files.
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef HIP_CONVERSION_ONNXTOHIP_UTILS_H
+#define HIP_CONVERSION_ONNXTOHIP_UTILS_H
+
+#include "hip/Conversion/OnnxToHip/Passes.h"
+#include "hip/Dialect/IR/HipDialect.h"
+#include "hip/Dialect/Transforms/Passes.h"
+
+#include "mlir/Dialect/Arith/IR/Arith.h"
+#include "mlir/Dialect/Bufferization/IR/Bufferization.h"
+#include "mlir/Dialect/Func/IR/FuncOps.h"
+#include "mlir/Dialect/MemRef/IR/MemRef.h"
+#include "mlir/Dialect/Tensor/IR/Tensor.h"
+#include "mlir/Dialect/Utils/ReshapeOpsUtils.h"
+#include "mlir/IR/BuiltinOps.h"
+#include "mlir/IR/PatternMatch.h"
+#include "mlir/Pass/Pass.h"
+#include "mlir/Transforms/GreedyPatternRewriteDriver.h"
+
+#include "llvm/ADT/STLExtras.h"
+#include "llvm/ADT/Sequence.h"
+#include "llvm/ADT/SmallSet.h"
+#include "llvm/Support/Debug.h"
+#include "llvm/Support/raw_ostream.h"
+
+#define DEBUG_TYPE "convert-onnx-to-hip"
+
+namespace mlir {
+namespace hip {
+
+//===----------------------------------------------------------------------===//
+// Helpers
+//===----------------------------------------------------------------------===//
+
+/// Sanitize an arbitrary string (typically an ONNX node name) into a valid
+/// MLIR bare identifier fragment.  Non-alphanumeric characters are replaced
+/// with '_', consecutive underscores are collapsed, and leading/trailing
+/// underscores are stripped.  Returns an empty string if the input yields
+/// no usable characters.
+inline std::string sanitizeForMlirIdentifier(llvm::StringRef raw) {
+  std::string sanitized;
+  sanitized.reserve(raw.size());
+  for (char c : raw) {
+    if (std::isalnum(static_cast<unsigned char>(c)) || c == '_')
+      sanitized.push_back(c);
+    else
+      sanitized.push_back('_');
+  }
+  // Collapse runs of underscores and trim leading/trailing ones.
+  std::string result;
+  result.reserve(sanitized.size());
+  bool lastWasUnderscore = true; // suppress leading '_'
+  for (char c : sanitized) {
+    if (c == '_') {
+      if (!lastWasUnderscore)
+        result.push_back(c);
+      lastWasUnderscore = true;
+    } else {
+      result.push_back(c);
+      lastWasUnderscore = false;
+    }
+  }
+  while (!result.empty() && result.back() == '_')
+    result.pop_back();
+  return result;
+}
+
+/// Create a tensor.empty for a DPS init operand.  Dynamic dimension sizes
+/// are extracted from \p source using tensor.dim at each dynamic index.
+/// Suitable for ops where the output shape aligns positionally with one input
+/// (e.g., softmax, element-wise).
+inline mlir::Value createEmptyTensor(mlir::OpBuilder &builder,
+                                     mlir::Location loc,
+                                     mlir::RankedTensorType resultType,
+                                     mlir::Value source) {
+  llvm::SmallVector<mlir::Value> dynSizes;
+  for (int64_t dimIdx : llvm::seq<int64_t>(resultType.getRank())) {
+    if (resultType.isDynamicDim(dimIdx))
+      dynSizes.push_back(
+          mlir::tensor::DimOp::create(builder, loc, source, dimIdx));
+  }
+  return mlir::tensor::EmptyOp::create(builder, loc, resultType.getShape(),
+                                       resultType.getElementType(), dynSizes);
+}
+
+/// Get !hip.context from function argument 0. Returns failure if the
+/// function has no arguments or the first argument is not !hip.context.
+inline mlir::FailureOr<mlir::Value>
+getContextArg(mlir::Operation *op, mlir::PatternRewriter &rewriter) {
+  auto funcOp = op->getParentOfType<mlir::func::FuncOp>();
+  if (!funcOp)
+    return rewriter.notifyMatchFailure(op, "not inside a function");
+  auto &entry = funcOp.getBody().front();
+  if (entry.getNumArguments() == 0)
+    return rewriter.notifyMatchFailure(op, "function has no arguments");
+  mlir::Value ctx = entry.getArgument(0);
+  if (!mlir::isa<mlir::hip::ContextType>(ctx.getType()))
+    return rewriter.notifyMatchFailure(op,
+                                       "first argument is not !hip.context");
+  return ctx;
+}
+
+// Pattern population functions (one per operator file)
+void populateMatMulConversionPatterns(RewritePatternSet &patterns,
+                                      MLIRContext *ctx);
+void populateTransposeConversionPatterns(RewritePatternSet &patterns,
+                                         MLIRContext *ctx);
+void populateElementwiseConversionPatterns(RewritePatternSet &patterns,
+                                           MLIRContext *ctx);
+void populateActivationConversionPatterns(RewritePatternSet &patterns,
+                                          MLIRContext *ctx);
+void populateCastConversionPatterns(RewritePatternSet &patterns,
+                                    MLIRContext *ctx);
+void populateReduceSumConversionPatterns(RewritePatternSet &patterns,
+                                         MLIRContext *ctx);
+void populateMatMulNBitsConversionPatterns(RewritePatternSet &patterns,
+                                           MLIRContext *ctx);
+void populateQMoEConversionPatterns(RewritePatternSet &patterns,
+                                    MLIRContext *ctx);
+void populateConvConversionPatterns(RewritePatternSet &patterns,
+                                    MLIRContext *ctx);
+void populateNormConversionPatterns(RewritePatternSet &patterns,
+                                    MLIRContext *ctx);
+void populateRotaryEmbeddingConversionPatterns(RewritePatternSet &patterns,
+                                               MLIRContext *ctx);
+void populateGqaConversionPatterns(RewritePatternSet &patterns,
+                                   MLIRContext *ctx);
+void populateGatherConversionPatterns(RewritePatternSet &patterns,
+                                      MLIRContext *ctx);
+void populateReshapeConversionPatterns(RewritePatternSet &patterns,
+                                       MLIRContext *ctx);
+
+} // namespace hip
+} // namespace mlir
+
+#endif // HIP_CONVERSION_ONNXTOHIP_UTILS_H
