+++ lib/Conversion/OnnxToHip/ReduceSumConversion.cpp
@@ -1,0 +1,102 @@
+/*
+ * Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+#include "OnnxToHipUtils.h"
+
+namespace mlir {
+namespace hip {
+namespace {
+
+/// onnx.ReduceSum -> hip.reduce_sum
+struct ReduceSumToHip : public mlir::RewritePattern {
+  ReduceSumToHip(mlir::MLIRContext *ctx)
+      : RewritePattern("onnx.ReduceSum", /*benefit=*/1, ctx) {}
+
+  mlir::LogicalResult
+  matchAndRewrite(mlir::Operation *op,
+                  mlir::PatternRewriter &rewriter) const override;
+};
+
+mlir::LogicalResult
+ReduceSumToHip::matchAndRewrite(mlir::Operation *op,
+                                mlir::PatternRewriter &rewriter) const {
+  auto ctxOrFailure = getContextArg(op, rewriter);
+  if (mlir::failed(ctxOrFailure))
+    return mlir::failure();
+  mlir::Value context = *ctxOrFailure;
+
+  mlir::Location loc = op->getLoc();
+  mlir::Value data = op->getOperand(0);
+  auto resultType =
+      mlir::cast<mlir::RankedTensorType>(op->getResult(0).getType());
+  mlir::Value init = createEmptyTensor(rewriter, loc, resultType, data);
+
+  // Extract noop_with_empty_axes attribute (defaults to 0 in ONNX)
+  int64_t noopWithEmptyAxes = 0;
+  if (auto noopAttr =
+          op->getAttrOfType<mlir::IntegerAttr>("noop_with_empty_axes")) {
+    noopWithEmptyAxes = noopAttr.getSInt();
+  }
+
+  // Handle axes: can be operand (opset 13+) or attribute (opset < 13)
+  // axes is always required in HIP dialect; create empty tensor<0xi64> when not
+  // provided
+  mlir::Value axesOperand;
+
+  if (op->getNumOperands() > 1) {
+    // Axes provided as operand (opset 13+)
+    axesOperand = op->getOperand(1);
+  } else {
+    // Axes provided as attribute - convert to constant tensor
+    llvm::SmallVector<int64_t> axesVec;
+    if (auto axesAttr = op->getAttrOfType<mlir::ArrayAttr>("axes")) {
+      for (auto a : axesAttr)
+        axesVec.push_back(
+            mlir::cast<mlir::IntegerAttr>(a).getValue().getSExtValue());
+    } else if (noopWithEmptyAxes == 0) {
+      // Default: reduce all axes (when noop_with_empty_axes is 0)
+      auto inputType = mlir::cast<mlir::RankedTensorType>(data.getType());
+      for (int64_t i : llvm::seq<int64_t>(inputType.getRank()))
+        axesVec.push_back(i);
+    } else {
+      // noop_with_empty_axes is 1 and no axes provided, axesVec remains empty
+      // (will create empty tensor<0xi64>)
+    }
+
+    // Create constant tensor for axes
+    auto axesType = mlir::RankedTensorType::get(
+        {static_cast<int64_t>(axesVec.size())}, rewriter.getI64Type());
+    auto axesAttr =
+        mlir::DenseIntElementsAttr::get(axesType, llvm::ArrayRef(axesVec));
+    axesOperand =
+        mlir::arith::ConstantOp::create(rewriter, loc, axesType, axesAttr);
+  }
+
+  // Extract keepdims attribute (defaults to 1 in ONNX)
+  int64_t keepdims = 1;
+  if (auto keepdimsAttr = op->getAttrOfType<mlir::IntegerAttr>("keepdims")) {
+    keepdims = keepdimsAttr.getSInt();
+  }
+
+  // Create hip.reduce_sum operation (axes always provided, may be empty)
+  auto keepdimsAttr = rewriter.getI64IntegerAttr(keepdims);
+  auto noopWithEmptyAxesAttr = rewriter.getI64IntegerAttr(noopWithEmptyAxes);
+  auto hipOp = mlir::hip::ReduceSumOp::create(
+      rewriter, loc, resultType, context, data, axesOperand, init, keepdimsAttr,
+      noopWithEmptyAxesAttr);
+
+  rewriter.replaceOp(op, hipOp->getResult(0));
+  return mlir::success();
+}
+
+} // namespace
+
+void mlir::hip::populateReduceSumConversionPatterns(
+    RewritePatternSet &patterns, MLIRContext *ctx) {
+  patterns.add<ReduceSumToHip>(ctx);
+}
+
+} // namespace hip
+} // namespace mlir
