+++ lib/Conversion/OnnxToHip/ReshapeConversion.cpp
@@ -1,0 +1,287 @@
+/*
+ * Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+#include "OnnxToHipUtils.h"
+
+namespace mlir {
+namespace hip {
+namespace {
+
+//===----------------------------------------------------------------------===//
+// Shape Operations Helpers (Reshape, Unsqueeze, Squeeze)
+//===----------------------------------------------------------------------===//
+
+/// Validate common requirements for Unsqueeze/Squeeze operations.
+/// Requires: 2 operands (data, axes), ranked tensors, matching element types,
+/// and constant axes (dynamic axes would require runtime shape computation).
+mlir::LogicalResult
+validateSqueezeUnsqueezeOp(mlir::Operation *op, mlir::PatternRewriter &rewriter,
+                           const char *tensorOpName, mlir::Value &data,
+                           mlir::Value &axes, mlir::RankedTensorType &inputType,
+                           mlir::RankedTensorType &outputType) {
+  if (op->getNumOperands() != 2)
+    return rewriter.notifyMatchFailure(op, "expected 2 operands (data, axes)");
+
+  data = op->getOperand(0);
+  axes = op->getOperand(1);
+
+  inputType = mlir::dyn_cast<mlir::RankedTensorType>(data.getType());
+  outputType =
+      mlir::dyn_cast<mlir::RankedTensorType>(op->getResult(0).getType());
+  if (!inputType || !outputType)
+    return rewriter.notifyMatchFailure(op, "expected ranked tensor types");
+  if (inputType.getElementType() != outputType.getElementType())
+    return rewriter.notifyMatchFailure(op, "element type mismatch");
+
+  auto axesDefOp = axes.getDefiningOp();
+  if (!axesDefOp) {
+    std::string msg =
+        "axes must be defined by a constant operation (block "
+        "argument not supported). Dynamic axes would require "
+        "runtime shape computation and cannot use zero-cost tensor.";
+    msg += tensorOpName;
+    return rewriter.notifyMatchFailure(op, msg);
+  }
+
+  bool isConstant = mlir::isa<mlir::arith::ConstantOp>(axesDefOp) ||
+                    axesDefOp->hasAttr("value");
+  if (!isConstant) {
+    std::string msg =
+        "axes must be constant (arith.constant or onnx.Constant). "
+        "Dynamic axes would require runtime shape computation and "
+        "cannot use zero-cost tensor.";
+    msg += tensorOpName;
+    msg += " approach";
+    return rewriter.notifyMatchFailure(op, msg);
+  }
+
+  return mlir::success();
+}
+
+/// Build output shape for expand_shape operations.
+/// Used by both Reshape and Unsqueeze when expanding dimensions.
+///
+/// For static dimensions: use compile-time size from outputType.
+/// For dynamic dimensions: extract from input via DimOp, dividing out any
+/// static dimensions in the same reassociation group.
+llvm::SmallVector<mlir::OpFoldResult> buildExpandShapeOutputShape(
+    mlir::PatternRewriter &rewriter, mlir::Location loc, mlir::Value data,
+    mlir::RankedTensorType outputType,
+    llvm::ArrayRef<mlir::ReassociationIndices> reassoc) {
+  int64_t outputRank = outputType.getRank();
+
+  llvm::SmallVector<int64_t> outDimToInDim(outputRank, -1);
+  for (auto [g, group] : llvm::enumerate(reassoc))
+    for (int64_t idx : group)
+      outDimToInDim[idx] = g;
+
+  llvm::SmallVector<mlir::OpFoldResult> outputShape;
+  for (int64_t i : llvm::seq<int64_t>(outputRank)) {
+    if (!outputType.isDynamicDim(i)) {
+      outputShape.push_back(rewriter.getIndexAttr(outputType.getDimSize(i)));
+      continue;
+    }
+
+    int64_t srcDim = outDimToInDim[i];
+    const auto &group = reassoc[srcDim];
+
+    int64_t staticProduct = 1;
+    for (int64_t idx : group)
+      if (!outputType.isDynamicDim(idx))
+        staticProduct *= outputType.getDimSize(idx);
+
+    mlir::Value inputSize =
+        mlir::tensor::DimOp::create(rewriter, loc, data, srcDim);
+    if (staticProduct == 1) {
+      outputShape.push_back(inputSize);
+    } else {
+      mlir::Value divisor =
+          mlir::arith::ConstantIndexOp::create(rewriter, loc, staticProduct);
+      mlir::Value dynSize =
+          mlir::arith::DivUIOp::create(rewriter, loc, inputSize, divisor);
+      outputShape.push_back(dynSize);
+    }
+  }
+
+  return outputShape;
+}
+
+//===----------------------------------------------------------------------===//
+// Reshape -> standard tensor ops (zero-cost metadata reinterpretation)
+//===----------------------------------------------------------------------===//
+
+/// onnx.Reshape -> tensor.expand_shape / tensor.collapse_shape.
+///
+/// Reshape is a zero-cost metadata operation: it reinterprets shape/strides
+/// without moving data.  We lower to standard MLIR tensor ops which
+/// bufferize to memref.expand_shape / memref.collapse_shape (zero-copy
+/// alias) and then to LLVM struct manipulation (same data pointer, new
+/// sizes/strides).  No HIP kernel is needed.
+struct ReshapeToStdTensor : public mlir::RewritePattern {
+  ReshapeToStdTensor(mlir::MLIRContext *ctx)
+      : RewritePattern("onnx.Reshape", /*benefit=*/1, ctx) {}
+
+  mlir::LogicalResult
+  matchAndRewrite(mlir::Operation *op,
+                  mlir::PatternRewriter &rewriter) const override {
+    mlir::Value data = op->getOperand(0);
+    auto inputType = mlir::dyn_cast<mlir::RankedTensorType>(data.getType());
+    auto outputType =
+        mlir::dyn_cast<mlir::RankedTensorType>(op->getResult(0).getType());
+    if (!inputType || !outputType)
+      return rewriter.notifyMatchFailure(op, "expected ranked tensor types");
+    if (inputType.getElementType() != outputType.getElementType())
+      return rewriter.notifyMatchFailure(op, "element type mismatch");
+
+    mlir::Location loc = op->getLoc();
+    int64_t inputRank = inputType.getRank();
+    int64_t outputRank = outputType.getRank();
+
+    // No-op: same type
+    if (inputType == outputType) {
+      rewriter.replaceOp(op, data);
+      return mlir::success();
+    }
+
+    // Different rank: expand or collapse
+    if (outputRank != inputRank) {
+      auto reassocOpt =
+          mlir::getReassociationIndicesForReshape(inputType, outputType);
+      if (!reassocOpt)
+        return rewriter.notifyMatchFailure(
+            op, "cannot compute reshape reassociation");
+
+      if (outputRank > inputRank) {
+        // Expand: use shared helper to build output shape
+        auto outputShape = buildExpandShapeOutputShape(rewriter, loc, data,
+                                                       outputType, *reassocOpt);
+        auto expandOp = mlir::tensor::ExpandShapeOp::create(
+            rewriter, loc, outputType, data, *reassocOpt, outputShape);
+        rewriter.replaceOp(op, expandOp.getResult());
+      } else {
+        // Collapse: no dynamic shape computation needed
+        auto collapseOp = mlir::tensor::CollapseShapeOp::create(
+            rewriter, loc, outputType, data, *reassocOpt);
+        rewriter.replaceOp(op, collapseOp.getResult());
+      }
+      return mlir::success();
+    }
+
+    // Same rank, different shape: collapse to 1-D then expand.
+    if (!inputType.hasStaticShape() || !outputType.hasStaticShape())
+      return rewriter.notifyMatchFailure(
+          op, "same-rank dynamic reshape not supported");
+
+    int64_t numElems = inputType.getNumElements();
+    if (numElems != outputType.getNumElements())
+      return rewriter.notifyMatchFailure(op, "element count mismatch");
+
+    auto flatType =
+        mlir::RankedTensorType::get({numElems}, inputType.getElementType());
+
+    mlir::ReassociationIndices allInputDims;
+    for (int64_t i : llvm::seq<int64_t>(inputRank))
+      allInputDims.push_back(i);
+    llvm::SmallVector<mlir::ReassociationIndices> collapseReassoc = {
+        allInputDims};
+    auto collapsed = mlir::tensor::CollapseShapeOp::create(
+        rewriter, loc, flatType, data, collapseReassoc);
+
+    mlir::ReassociationIndices allOutputDims;
+    for (int64_t i : llvm::seq<int64_t>(outputRank))
+      allOutputDims.push_back(i);
+    llvm::SmallVector<mlir::ReassociationIndices> expandReassoc = {
+        allOutputDims};
+    llvm::SmallVector<mlir::OpFoldResult> flatOutputShape;
+    for (int64_t i : llvm::seq<int64_t>(outputRank))
+      flatOutputShape.push_back(
+          rewriter.getIndexAttr(outputType.getDimSize(i)));
+    auto expanded = mlir::tensor::ExpandShapeOp::create(
+        rewriter, loc, outputType, collapsed.getResult(), expandReassoc,
+        flatOutputShape);
+
+    rewriter.replaceOp(op, expanded.getResult());
+    return mlir::success();
+  }
+};
+
+//===----------------------------------------------------------------------===//
+// Unsqueeze -> standard tensor ops (zero-cost metadata reinterpretation)
+//===----------------------------------------------------------------------===//
+
+/// onnx.Unsqueeze -> tensor.expand_shape (zero-cost metadata operation).
+struct UnsqueezeToStdTensor : public mlir::RewritePattern {
+  UnsqueezeToStdTensor(mlir::MLIRContext *ctx)
+      : RewritePattern("onnx.Unsqueeze", /*benefit=*/1, ctx) {}
+
+  mlir::LogicalResult
+  matchAndRewrite(mlir::Operation *op,
+                  mlir::PatternRewriter &rewriter) const override {
+    mlir::Value data, axes;
+    mlir::RankedTensorType inputType, outputType;
+    if (auto result = validateSqueezeUnsqueezeOp(
+            op, rewriter, "expand_shape", data, axes, inputType, outputType);
+        failed(result))
+      return result;
+
+    auto reassocOpt =
+        mlir::getReassociationIndicesForReshape(inputType, outputType);
+    if (!reassocOpt)
+      return rewriter.notifyMatchFailure(
+          op, "cannot compute unsqueeze reassociation");
+
+    mlir::Location loc = op->getLoc();
+    auto outputShape = buildExpandShapeOutputShape(rewriter, loc, data,
+                                                   outputType, *reassocOpt);
+    auto expandOp = mlir::tensor::ExpandShapeOp::create(
+        rewriter, loc, outputType, data, *reassocOpt, outputShape);
+    rewriter.replaceOp(op, expandOp.getResult());
+    return mlir::success();
+  }
+};
+
+//===----------------------------------------------------------------------===//
+// Squeeze -> standard tensor ops (zero-cost metadata reinterpretation)
+//===----------------------------------------------------------------------===//
+
+/// onnx.Squeeze -> tensor.collapse_shape (zero-cost metadata operation).
+struct SqueezeToStdTensor : public mlir::RewritePattern {
+  SqueezeToStdTensor(mlir::MLIRContext *ctx)
+      : RewritePattern("onnx.Squeeze", /*benefit=*/1, ctx) {}
+
+  mlir::LogicalResult
+  matchAndRewrite(mlir::Operation *op,
+                  mlir::PatternRewriter &rewriter) const override {
+    mlir::Value data, axes;
+    mlir::RankedTensorType inputType, outputType;
+    if (auto result = validateSqueezeUnsqueezeOp(
+            op, rewriter, "collapse_shape", data, axes, inputType, outputType);
+        failed(result))
+      return result;
+
+    auto reassocOpt =
+        mlir::getReassociationIndicesForReshape(inputType, outputType);
+    if (!reassocOpt)
+      return rewriter.notifyMatchFailure(
+          op, "cannot compute squeeze reassociation");
+
+    mlir::Location loc = op->getLoc();
+    auto collapseOp = mlir::tensor::CollapseShapeOp::create(
+        rewriter, loc, outputType, data, *reassocOpt);
+    rewriter.replaceOp(op, collapseOp.getResult());
+    return mlir::success();
+  }
+};
+
+} // namespace
+
+void mlir::hip::populateReshapeConversionPatterns(RewritePatternSet &patterns,
+                                                  MLIRContext *ctx) {
+  patterns.add<ReshapeToStdTensor, UnsqueezeToStdTensor, SqueezeToStdTensor>(
+      ctx);
+}
+
+} // namespace hip
+} // namespace mlir
