+++ lib/Conversion/OnnxToHip/RotaryEmbeddingConversion.cpp
@@ -1,0 +1,132 @@
+/*
+ * Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+#include "OnnxToHipUtils.h"
+
+namespace mlir {
+namespace hip {
+namespace {
+
+/// onnx.Custom(RotaryEmbedding) -> hip.rope
+struct RotaryEmbeddingToHip : public mlir::RewritePattern {
+  RotaryEmbeddingToHip(mlir::MLIRContext *ctx)
+      : RewritePattern("onnx.Custom", /*benefit=*/1, ctx) {}
+
+  mlir::LogicalResult
+  matchAndRewrite(mlir::Operation *op,
+                  mlir::PatternRewriter &rewriter) const override;
+};
+
+mlir::LogicalResult
+RotaryEmbeddingToHip::matchAndRewrite(mlir::Operation *op,
+                                      mlir::PatternRewriter &rewriter) const {
+  // Check if this is RotaryEmbedding
+  auto funcNameAttr = op->getAttrOfType<mlir::StringAttr>("function_name");
+  if (!funcNameAttr || funcNameAttr.getValue() != "RotaryEmbedding")
+    return rewriter.notifyMatchFailure(op, "not a RotaryEmbedding operation");
+
+  // Check domain is "com.microsoft"
+  auto domainAttr = op->getAttrOfType<mlir::StringAttr>("domain_name");
+  if (!domainAttr || domainAttr.getValue() != "com.microsoft")
+    return rewriter.notifyMatchFailure(
+        op, "domain must be com.microsoft for RotaryEmbedding");
+
+  auto ctxOrFailure = getContextArg(op, rewriter);
+  if (mlir::failed(ctxOrFailure))
+    return rewriter.notifyMatchFailure(op, "missing context argument");
+  mlir::Value context = *ctxOrFailure;
+
+  mlir::Location loc = op->getLoc();
+
+  // Check operands (should be 4: input, position_ids, cos_cache, sin_cache)
+  if (op->getNumOperands() != 4)
+    return rewriter.notifyMatchFailure(
+        op, "expected 4 operands for RotaryEmbedding");
+
+  mlir::Value input = op->getOperand(0);
+  mlir::Value positionIds = op->getOperand(1);
+  mlir::Value cosCache = op->getOperand(2);
+  mlir::Value sinCache = op->getOperand(3);
+
+  // Extract attributes
+  auto interleavedAttr = op->getAttrOfType<mlir::IntegerAttr>("interleaved");
+  if (!interleavedAttr)
+    return rewriter.notifyMatchFailure(op, "missing interleaved attribute");
+
+  auto numHeadsAttr = op->getAttrOfType<mlir::IntegerAttr>("num_heads");
+  if (!numHeadsAttr)
+    return rewriter.notifyMatchFailure(op, "missing num_heads attribute");
+
+  auto rotaryDimAttr =
+      op->getAttrOfType<mlir::IntegerAttr>("rotary_embedding_dim");
+  if (!rotaryDimAttr)
+    return rewriter.notifyMatchFailure(
+        op, "missing rotary_embedding_dim attribute");
+
+  int64_t numHeadsVal = numHeadsAttr.getSInt();
+  int64_t rotaryDimVal = rotaryDimAttr.getSInt();
+
+  // ONNX com.microsoft.RotaryEmbedding: 0 means "infer from tensor shapes"
+  //   cos_cache: [max_seq, rotary_dim/2] -> rotary_dim = last_dim * 2
+  //   input:     [batch, seq, hidden]     -> num_heads = hidden / rotary_dim
+  if (rotaryDimVal == 0) {
+    auto cosCacheType =
+        mlir::dyn_cast<mlir::RankedTensorType>(cosCache.getType());
+    if (cosCacheType && cosCacheType.hasStaticShape() &&
+        cosCacheType.getRank() >= 2) {
+      rotaryDimVal = cosCacheType.getShape().back() * 2;
+    } else {
+      return rewriter.notifyMatchFailure(
+          op, "Cannot infer rotary_embedding_dim: "
+              "cos_cache must have static shape with rank >= 2");
+    }
+  }
+
+  if (numHeadsVal == 0 && rotaryDimVal > 0) {
+    auto inputType = mlir::dyn_cast<mlir::RankedTensorType>(input.getType());
+    if (inputType && inputType.hasStaticShape() && inputType.getRank() >= 1) {
+      int64_t hidden = inputType.getShape().back();
+      numHeadsVal = hidden / rotaryDimVal;
+    } else {
+      return rewriter.notifyMatchFailure(op, "Cannot infer num_heads: "
+                                             "input must have static shape");
+    }
+  }
+
+  // Convert to i64 attributes (using inferred or original values)
+  auto interleavedI64Attr =
+      rewriter.getI64IntegerAttr(interleavedAttr.getSInt());
+  auto numHeadsI64Attr = rewriter.getI64IntegerAttr(numHeadsVal);
+  auto rotaryDimI64Attr = rewriter.getI64IntegerAttr(rotaryDimVal);
+
+  // Should have 1 result: output
+  if (op->getNumResults() != 1)
+    return rewriter.notifyMatchFailure(op,
+                                       "expected 1 result for RotaryEmbedding");
+
+  auto resultType =
+      mlir::cast<mlir::RankedTensorType>(op->getResult(0).getType());
+
+  // Create init tensor
+  mlir::Value init = createEmptyTensor(rewriter, loc, resultType, input);
+
+  // Create hip.rope operation
+  auto hipOp = mlir::hip::RopeOp::create(
+      rewriter, loc, resultType, context, input, positionIds, cosCache,
+      sinCache, init, interleavedI64Attr, numHeadsI64Attr, rotaryDimI64Attr);
+
+  rewriter.replaceOp(op, hipOp->getResult(0));
+  return mlir::success();
+}
+
+} // namespace
+
+void mlir::hip::populateRotaryEmbeddingConversionPatterns(
+    RewritePatternSet &patterns, MLIRContext *ctx) {
+  patterns.add<RotaryEmbeddingToHip>(ctx);
+}
+
+} // namespace hip
+} // namespace mlir
